changed doube interface to bigdecimal, but not in RPN - it stay calculate double

+++ src/main/java/com/fedorovigor/calculator/model/dto/ExpressionDto.java
@@ -2,16 +2,17 @@ package com.fedorovigor.calculator.model.dto;
 
 import com.fedorovigor.calculator.model.entity.ExpressionEntity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ExpressionDto {
 
 
     private final String expression;
-    private final Double result;
+    private final BigDecimal result;
     private final LocalDateTime saveTime;
 
-    public ExpressionDto(String expression, Double result, LocalDateTime saveTime) {
+    public ExpressionDto(String expression, BigDecimal result, LocalDateTime saveTime) {
         this.saveTime = saveTime;
         this.expression = expression;
         this.result = result;
@@ -27,7 +28,7 @@ public class ExpressionDto {
         return expression;
     }
 
-    public Double getResult() {
+    public BigDecimal getResult() {
         return result;
     }
