+++ src/main/java/com/fedorovigor/calculator/service/CalculatorService.java
@@ -2,11 +2,12 @@ package com.fedorovigor.calculator.service;
 
 import com.fedorovigor.calculator.model.dto.ExpressionDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CalculatorService {
 
-    Double calculateAndSave(String expr);
+    BigDecimal calculateAndSave(String expr);
     List<ExpressionDto> getLasTen();
 
 }
